Huge change! All prisoner, circles and locations variables changed to class attributes! Many bug fixes. Now need to complete the todo list, then push.

+++ UI_server.py
@@ -14,7 +14,8 @@ MINUTES_GAP_CHECK = 1
 
 
 def get_prisoner_data_and_current_location_and_red_circles(content, db):
-    try:
+    #try:
+    if True:
         """
         data_to_return:
             0: ID
@@ -26,158 +27,179 @@ def get_prisoner_data_and_current_location_and_red_circles(content, db):
             6: red_circles[[circle_id, prisoner_id, radius, lat, lng, type], [], []...]
         """
         prisoner_id = content["input"]
-        try:
-            prisoner_personal_data = db.get_prisoner_personal_data(prisoner_id)[0]
-            prisoner_location = db.get_prisoner_locations(prisoner_id)[-1]
+        #try:
+        if True:
+            prisoner_data = db.get_prisoner_personal_data(prisoner_id)
+            prisoner_last_location = db.get_prisoner_last_location(prisoner_id)
             prisoner_red_circles = db.get_all_prisoner_red_circles(prisoner_id)
-            data_to_return = [prisoner_id, prisoner_personal_data[1], prisoner_personal_data[2],
-                              [prisoner_location[3], prisoner_location[4]], prisoner_location[2],
-                              prisoner_personal_data[3], prisoner_red_circles]
-        except IndexError:
+            if prisoner_last_location is not None and prisoner_data is not None and prisoner_red_circles is not None:
+                prisoner_red_circles_ready = []
+                for circle in prisoner_red_circles:
+                    prisoner_red_circles_ready.append(circle.get_circle_listed())
+                data_to_return = [prisoner_id, prisoner_data.get_name(), prisoner_data.get_national_identifier(),
+                                  [prisoner_last_location.get_lat(), prisoner_last_location.get_lng()],
+                                  prisoner_last_location.get_date(), prisoner_data.get_status(), prisoner_red_circles_ready]
+                return json.dumps(data_to_return)
+            else:
+                return "No prisoner in the system"
+
+        #except IndexError:
             return "No Location Measured Yet"
-        return json.dumps(data_to_return)
-    except:
+    #except:
+        # TODO i got error "submit error", why?
         return "Submit Prisoner Error"
 
 
+# return list of all prisoners
+def get_all_prisoners(db):
+    #try:
+    if True:
+        prisoners_data = db.get_all_prisoners()
+        ret_list = []
+        for prisoner_ret in prisoners_data:
+            listed_prisoner = [prisoner_ret.get_prisoner_id(), prisoner_ret.get_name(),
+                               prisoner_ret.get_national_identifier(), prisoner_ret.get_status()]
+            ret_list.append(listed_prisoner)
+        return json.dumps(ret_list)
+    #except:
+        return "LoadingAllPrisonersForListError"
+
+
 def get_all_problems_and_new_alerts(db):
-    # data place holders for time differences
-    last_location = -1
-    time_stamp_place = 2
-    # data place holders for red circles
-    radius_placer = 2
-    location_placer_lat = 3
-    location_placer_lng = 4
-    circle_type_placer = 5
-    # data place holders for prisoner`s location data
-    prisoner_location_placer_lat = 3
-    prisoner_location_placer_lng = 4
-    try:
+    #try:
+    if True:
         all_new_log_data = []
         # running through all prisoners
         for prisoner_data in db.get_all_prisoners():
             # now, we will check if the prisoner is in unwanted place
 
-            prisoner_id = prisoner_data[0]
+            prisoner_id = prisoner_data.get_prisoner_id()
 
             # getting all the prisoner`s red circles
             all_red_circles = db.get_all_prisoner_red_circles(prisoner_id)
-            prisoner_last_location = db.get_prisoner_personal_data(prisoner_id)
+            prisoner_last_location = db.get_prisoner_last_location(prisoner_id)
+            # TODO prisoner_last_location might be None (no location measured yet)
 
-            # getting the prisoner`s last timestamp and current timestamp
-            last_timestamp = db.get_prisoner_locations(prisoner_id)[last_location][time_stamp_place]
-            current_time = str(datetime.now())
-            time_format = '%Y-%m-%d %H:%M:%S.%f'
-            prisoner_time_delta = datetime.strptime(current_time, time_format) - datetime.strptime(last_timestamp, time_format)
+            if prisoner_last_location is not None:
+                # getting the prisoner`s last timestamp and current timestamp
+                last_timestamp = prisoner_last_location.get_date()
+                current_time = str(datetime.now())
+                time_format = '%Y-%m-%d %H:%M:%S.%f'
+                prisoner_time_delta = datetime.strptime(current_time, time_format) - datetime.strptime(last_timestamp,
+                                                                                                       time_format)
 
-            # the problem checker - if bigger than 0 - problem
-            is_problem = 0
+                # the problem checker - if bigger than 0 - problem
+                is_problem = 0
 
-            # running through all the red circles
-            for red_circle in all_red_circles:
-                circle_position = (red_circle[location_placer_lat], red_circle[location_placer_lng])
-                prisoner_position = (prisoner_last_location[prisoner_location_placer_lat], prisoner_last_location[prisoner_location_placer_lng])
-                if Distance.coordinate_dis(circle_position, prisoner_position) <= red_circle[radius_placer]:
-                    # the prisoner is in the circle
-                    if red_circle[circle_type_placer] == "1":
-                        # Not allowed in the circle - Problem
-                        is_problem += 1
+                # running through all the red circles
+                for red_circle in all_red_circles:
+                    circle_position = (red_circle.get_lat(), red_circle.get_lng())
+                    prisoner_position = (prisoner_last_location.get_lat(), prisoner_last_location.get_lng())
+                    if Distance.coordinate_dis(circle_position, prisoner_position) <= red_circle.get_radius():
+                        # the prisoner is in the circle
+                        if red_circle.get_circle_type() == "1":
+                            # Not allowed in the circle - Problem
+                            is_problem += 1
+                        else:
+                            # allowed only in the circle - Good
+                            is_problem += 0
                     else:
-                        # allowed only in the circle - Good
-                        is_problem += 0
+                        # the prisoner is out of the circle
+                        if red_circle.get_circle_type() == "1":
+                            # Not allowed in the circle - Good
+                            is_problem += 0
+                        else:
+                            # allowed only in the circle - Problem
+                            is_problem += 1
+
+                # checking if the client didn't sent location for a while
+                if prisoner_time_delta > timedelta(minutes=MINUTES_GAP_CHECK):
+                    is_problem += 1
                 else:
-                    # the prisoner is out of the circle
-                    if red_circle[circle_type_placer] == "1":
-                        # Not allowed in the circle - Good
-                        is_problem += 0
-                    else:
-                        # allowed only in the circle - Problem
-                        is_problem += 1
+                    is_problem += 0
 
-            # checking if the client didn't sent location for a while
-            if prisoner_time_delta > timedelta(minutes=MINUTES_GAP_CHECK):
-                is_problem += 1
+                # finally - checking the problem checker
+                if is_problem >= 1:
+                    db.change_prisoner_status(prisoner_id, new_status=1)
+                else:
+                    db.change_prisoner_status(prisoner_id, new_status=0)
+                # adding the current prisoner to the list
+                all_new_log_data.append(prisoner_data.get_prisoner_listed())
+
             else:
-                is_problem += 0
-
-            # finally - checking the problem checker
-            if is_problem >= 1:
                 db.change_prisoner_status(prisoner_id, new_status=1)
-            else:
-                db.change_prisoner_status(prisoner_id, new_status=0)
-
-            # adding the current prisoner to the list
-            all_new_log_data.append(list(prisoner_data))
+                all_new_log_data.append(prisoner_data.get_prisoner_listed())
 
         return json.dumps(all_new_log_data)
-    except:
+    #except:
         return "LogError"
 
 
 def add_or_remove_red_circle(content, db):
-    try:
+    #try:
+    if True:
         data = json.loads(content["input"])
-        prisoner_id = data[0]
-        lat = data[1]
-        lng = data[2]
+        # opening data and creating RedCircle
         add_remove_type = data[3]
-        radius = data[4]
-        circle_type = data[5]
+        prisoner_data = DB.RedCircle(None, prisoner_id=data[0], radius=data[4], lat=data[1], lng=data[2],
+                                     circle_type=data[5])
+
         # function - adding
         if add_remove_type == "ADD":
             # checking radius
-            if int(radius) > 0:
-                db.add_red_circle(prisoner_id, radius, lat, lng, circle_type)
+            if int(prisoner_data.get_radius()) > 0:
+                db.add_red_circle(prisoner_data.get_prisoner_id(), prisoner_data.get_radius(), prisoner_data.get_lat(),
+                                  prisoner_data.get_lng(), prisoner_data.get_circle_type())
                 return "Added Circle"
             else:
                 return "Radius Can`t be 0 or negative! (Error)"
         # function - removing
         elif add_remove_type == "REMOVE":
             try:
-                db.remove_red_circle(prisoner_id, lat, lng)
+                db.remove_red_circle(prisoner_data.get_prisoner_id(), prisoner_data.get_lat(), prisoner_data.get_lng())
                 return "Removed Circle"
             except:
                 return "Data Base Error!"
         else:
             return "No Such Command! (Error)"
-    except:
+    #except:
         return "Value Error!"
 
 
 def new_prisoner(content, db):
-    try:
+    #try:
+    if True:
         data = json.loads(content["input"])
-        prisoner_national_id = data[0]
-        prisoner_name = data[1]
-        db.insert_new_prisoner(prisoner_id=prisoner_national_id, name=prisoner_name,
-                               national_identifier=prisoner_national_id, status=False)
+        prisoner_data = DB.Prisoner(data[0], data[1], data[0], False)
+        db.insert_new_prisoner(prisoner_data, status=False)
         return "Good"
-    except:
+    #except:
         return "ValueError"
 
 
 def prisoner_new_location(content, db):
-    try:
+    #try:
+    if True:
         data = json.loads(content["input"])
-        # loading the data that received
-        client_national_id = data[0]
-        location_lat = data[1]
-        location_lng = data[2]
         # creating timestamp
         date = datetime.now()
+        # loading the data that received
+        location_data = DB.Location(None, data[0], date, data[1], data[2])
+
         # inserting the new location into the database
-        db.insert_new_location(client_national_id, date, location_lat, location_lng)
+        db.insert_new_location(location_data)
         return "Good"
-    except:
+    #except:
         return "LocationError"
 
 
 def prisoner_unknown_location(content, db):
     prisoner_id = json.loads(content["input"])
-    try:
+    #try:
+    if True:
         db.change_prisoner_status(prisoner_id, new_status=1)
         return "GettingPrisonerLocationError"
-    except:
+    #except:
         print("Inner error. please call the code manager")
         return "GettingPrisonerLocationErrorAndDataBaseError"
 
@@ -190,7 +212,8 @@ def get_text_from_file(file_name):
 
 def get_help_text(content):
     help_type = content["input"]
-    try:
+    #try:
+    if True:
         if help_type == "for_users":
             return get_text_from_file("help_for_users.txt")
         elif help_type == "for_admins":
@@ -201,16 +224,17 @@ def get_help_text(content):
             return get_text_from_file("common_f_and_q.txt")
         else:
             return json.dumps("CommandNotFound")
-    except:
+    #except:
         return json.dumps("TextFileProblem")
 
 
 # removing prisoner
 def remove_prisoner(content, db):
-    try:
+    #try:
+    if True:
         db.remove_prisoner(content["input"])
         return "Prisoner Removed Successfully"
-    except:
+    #except:
         return "Prisoner Not Found"
 
 
@@ -222,10 +246,7 @@ def commend_checker(content, db):
 
     # command type: sending back all prisoners
     elif content["type"] == "get_all_prisoners_id_and_name_ect":
-        try:
-            return json.dumps(db.get_all_prisoners())
-        except:
-            return "LoadingAllPrisonersForListError"
+        return get_all_prisoners(db)
 
     # command type: get all new things that happened and all problems (status==1)
     elif content["type"] == "get_all_problems_and_new_alerts":
